+++ src/Jellyfin.Plugin.ListenBrainz/Tasks/SyncWeeklyPlaylists/WeeklyPlaylistCandidate.cs
@@ -1,5 +1,5 @@
 using Jellyfin.Plugin.ListenBrainz.Api.Models;
 
-namespace Jellyfin.Plugin.ListenBrainz.Tasks;
+namespace Jellyfin.Plugin.ListenBrainz.Tasks.SyncWeeklyPlaylists;
 
 internal sealed record WeeklyPlaylistCandidate(Playlist Playlist, WeeklyPlaylistType Type);
